Better handle symlinks with destructive system calls.
The unlink, rmdir, and rename system calls behave unexpectedly
if you resolve the symbolic link and perform the operation on that
instead.  It would leave a dangling symlink and delete the original
instead which is probably not the intention.

This commit adds another parameter to process_path to enable or disable
the symlink resolution.  The bindfs_unlink, bindfs_rmdir and
bindfs_rename functions use this parameter to produce the expected
result.  rm -rf will still behave differently, deleting the contents of
the symlinked directory before removing the symlink and not the
linked directory.