tests: fix building out of place
The path to the source files should be top_srcdir not top_builddir. When
building in place these are both the same so this doesn't fail but is
technically wrong. This change will allow both cases to work.

To demonstrace how this was broken before and to ensure it is fixed now
you can do the following:
./autogen.sh && mkdir -p build && cd build && ../configure && make

Running make check out of place still doesn't work but that will require
a bit more change.