tests: Fix run make check out of place
When building out of the place the executables are not in the same
location as the scripts being called. This means that the test scripts
need to be able to handle the executables being the in the current
directory and the script/source files being in a different directory.

This will work now work with both building in place and building out of
place and can be tested as follows:
/autogen.sh && mkdir -p build && cd build && ../configure && make check